renamed rest-api folder to restapi
