+++ src/restapi/content/delete.ts
